fix:加固自动化设备与图像识别边界

+++ src/AutoSwshRng.Upstream/EasyConCaptureServices.cs
@@ -117,8 +117,8 @@ public sealed class EasyConImageRecognitionService : IImageRecognitionService
         {
             if (request.Region is { } region)
             {
-                if (region.X + region.Width > source.Width
-                    || region.Y + region.Height > source.Height)
+                if ((long)region.X + region.Width > source.Width
+                    || (long)region.Y + region.Height > source.Height)
                 {
                     throw new UpstreamOperationException(
                         UpstreamErrorCode.Validation,
